fix(color-picker): slider 初始化 thumb 位置计算问题
